refactor(smappshot): extract signing utils and add Tenant enum
- Move signing/URL utilities to signing.go
- Add Tenant type with TenantBalyIQ and TenantBalyLBN constants
- WithTenant now accepts Tenant instead of string

AI-Assisted: true
AI-Model: claude-sonnet-4-6
